pipeline/requirements.txt
modal
requests
mercantile
pillow
